+++ config.py
@@ -1,0 +1,7 @@
+# config.py: เก็บการตั้งค่าทั้งหมดของแอป
+import os
+
+class Config:
+    SECRET_KEY = os.environ.get('SECRET_KEY', 'mykey')
+    SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL', 'sqlite:///database.db')
+    SQLALCHEMY_TRACK_MODIFICATIONS = False
